% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem.lint-package.R
\docType{package}
\name{golem.lint-package}
\alias{golem.lint}
\alias{golem.lint-package}
\title{golem.lint: Enhancing Code Quality for Golem Applications}
\description{
The `golem.lint` package provides linting tools for building `Shiny` applications. It helps developers identify and correct potential issues in their code, ensuring adherence to best practices and improving overall code quality.
}
\author{
\strong{Maintainer}: Colin Fay \email{contact@colinfay.me} (\href{https://orcid.org/0000-0001-7343-1846}{ORCID})

Authors:
\itemize{
  \item Arthur Bréant \email{arthur@thinkr.fr}
}

Other contributors:
\itemize{
  \item ThinkR [copyright holder]
}

}
\keyword{internal}
